Install twonky and integrate with systemd startup and control
This script will install Twonky 8.5.x to /opt/twonky, with working files in /var/opt/twonky, logs in /var/log/twonky, and configuration in /etc/twonky. 

A service account is created and used to manage access to working files and directories, the systemd unit file is copied to /etc/systemd/system, enabled on startup, and then the twonky service is started.